completely exclude build folder from version control
Note: Although this entry was already in the .gitignore file, I believe that entry was added when "build" folder had already been created. Therefore, since .gitignore cannot affect files that is already in version control, such entry was in fact useless.

To fix it, I temporarily commented out "build" entry in .gitignore, removed the folder locally and from remote, and added that entry back.

Now as long as we delete "build" folder locally, we can get rid of it forever.